NCL-5723 Remove unused code and get tests working
